First commit
Solved the data validation problem using error checking with try/except clauses. The program attempts to convert each string to int, and then checks if it is a positive int. The number is then added to a new list. If the cast to int fails, the except fires, continuing the loop.